backend/requirements.txt
Flask
flask-cors
python-dotenv
requests
scikit-learn
pandas
numpy
